requirements.txt
uvicorn==0.22.0
fastapi==0.100.0
pydantic[email]==2.0.3
python-decouple==3.8
pymongo==4.7.3
pyjwt==2.8.0
pytz==2024.1